requirements.txt
aspose-slides
openai
anthropic
streamlit
Pillow
